+++ composeunstyled-tooltip/src/androidInstrumentedTest/kotlin/Tooltip.androidTest.kt
@@ -44,7 +44,7 @@ class TooltipTest {
 
   private fun ComposeUiTest.setPaddedContent(content: @Composable () -> Unit) {
     setContent {
-      PortalHost {
+      TooltipHost {
         Box(Modifier.padding(100.dp)) {
           content()
         }